fix loss of information due to quality variable type declaration in OF ABI

+++ sw/airborne/modules/ctrl/optical_flow_landing.c
@@ -118,7 +118,7 @@ static abi_event optical_flow_ev;
 /// Callback function of the ground altitude
 static void vertical_ctrl_agl_cb(uint8_t sender_id __attribute__((unused)), float distance);
 // Callback function of the optical flow estimate:
-static void vertical_ctrl_optical_flow_cb(uint8_t sender_id __attribute__((unused)), uint32_t stamp, int16_t flow_x, int16_t flow_y, int16_t flow_der_x, int16_t flow_der_y, uint8_t quality, float size_divergence, float dist);
+static void vertical_ctrl_optical_flow_cb(uint8_t sender_id __attribute__((unused)), uint32_t stamp, int16_t flow_x, int16_t flow_y, int16_t flow_der_x, int16_t flow_der_y, float quality, float size_divergence, float dist);
 
 struct OpticalFlowLanding of_landing_ctrl;
 
@@ -480,7 +480,7 @@ static void vertical_ctrl_agl_cb(uint8_t sender_id, float distance)
 {
   of_landing_ctrl.agl = distance;
 }
-static void vertical_ctrl_optical_flow_cb(uint8_t sender_id, uint32_t stamp, int16_t flow_x, int16_t flow_y, int16_t flow_der_x, int16_t flow_der_y, uint8_t quality, float size_divergence, float dist)
+static void vertical_ctrl_optical_flow_cb(uint8_t sender_id, uint32_t stamp, int16_t flow_x, int16_t flow_y, int16_t flow_der_x, int16_t flow_der_y, float quality, float size_divergence, float dist)
 {
   divergence_vision = size_divergence;
   vision_message_nr++;
